Send a workspace change event when the workspace is created.
Going back on my decision to remove this new behavior. In order for
the BlocklyWorkspace component to be useful, the component needs to be
able to send an event when an instance of the workspace is created.
This way, additional workspace configuration can be done outside of
the component (e.g: Attaching a built-in event handler, like
Blockly.Events.disableOrphans)

I am implementing it slightly differently this time, using a boolean
state to keep track if the initial workspace creation has invoked a
workspace changed event.